start merging recording and transcribing...
